transcription/requirements.txt
fastapi
uvicorn
python-multipart
pyaudio 
numpy
openai-whisper
torch 
torchvision 
torchaudio
pydantic
redis
paho-mqtt